requirements.txt
numpy>=1.21.0
pandas>=1.3.0
rasterio>=1.2.0
geopandas>=0.9.0
shapely>=1.7.0
sentinelsat>=1.1.0
earthengine-api>=0.1.0
geemap>=0.8.0
laspy>=2.0.0
matplotlib>=3.4.0
plotly>=5.1.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
scipy>=1.7.0
requests>=2.26.0
tqdm>=4.62.0
python-dateutil>=2.8.0
pytz>=2021.1
folium>=0.12.0
openai>=0.27.0
opencv-python>=4.5.0
pillow>=8.3.0
seaborn>=0.11.0
python-pptx>=0.6.21 